Ugly Form

+++ src/main/java/Phonebook/PhoneBook.java
@@ -5,10 +5,10 @@ import java.io.Serializable;
 public class PhoneBook implements Serializable {
   private Integer mId;
   private String mName;
-  private Integer mPhone;
+  private Long mPhone;
   private String mEmail;
 
-  public PhoneBook(String name, Integer phonenumber, String email) {
+  public PhoneBook(String name, Long phonenumber, String email) {
     mName = name;
     mPhone = phonenumber;
     mEmail = email;
@@ -25,11 +25,11 @@ public class PhoneBook implements Serializable {
     return mName;
   }
 
-  public void setPhone(Integer phone) {
+  public void setPhone(Long phone) {
     mPhone = phone;
   }
 
-  public Integer getPhone() {
+  public Long getPhone() {
     return mPhone;
   }
